Calculate sign separately from formatting
We had 3 root causes to the wrongly sign behaviour:
1. The input was truncated before calculate the sign, thus
   numbers between 0 and -1 were wrongly formatted (wrong sign).
2. The calculation of the decimal part considered the sign of the input.
   This caused a rounding bug for negative decimal parts. For example formatting
   -0.00099 with 3 decimal places would round the 99 up to 1, which resulted in
   "0.0-1"
3. Finally we could not calculate the sign before we finish formatting the
   number, because the truncated result could change the final sign.
   For example the sign of "-0.0001" is "-" but if we are formatting with only
   2 decimals the result should be "0.00" and not "-0.00"

The solution of the first and third problems were to extract the sign
calculation to the helpers and call it when assembling the final result.
To accomplish this we needed to calculate the decimals before the assemble, and
not in place as we were doing before.

The solution of the second problem is simply to use the absolute of the number
to calculate the decimal part of the result.