Fix incorrect regex in FormatNumber.Locales.fromString
The regex is supposed to extract all non-digit characters from the string, but its usage with Regex.replace
actually led to it extracting all digits - this broke the rest of the parsing function and resulted in
an incorrect locale.

In addition, the FormatNumber.Locales module wasn't being verified by the tests, so this error was missed.